test(button): migrate strong
