+++ core/src/components/button/test/strong/button.e2e.ts-snapshots/button-strong-in-buttons-ios-ltr-Mobile-Firefox-linux.png
